More warnings gone.

+++ src/doom/deh_cheat.c
@@ -77,7 +77,7 @@ static void DEH_CheatParseLine(deh_context_t *context, char *line, void *tag)
     char *variable_name;
     char *value;
     unsigned char *unsvalue;
-    unsigned int i;
+    int32_t i;
 
     if (!DEH_ParseAssignment(line, &variable_name, &value))
     {
@@ -106,7 +106,7 @@ static void DEH_CheatParseLine(deh_context_t *context, char *line, void *tag)
         // If the cheat length exceeds the Vanilla limit, stop.  This
         // does not apply if we have the limit turned off.
 
-        if (!deh_allow_long_cheats && i >= cheat->seq->sequence_len)
+        if (!deh_allow_long_cheats && (size_t)i >= cheat->seq->sequence_len)
         {
             DEH_Warning(context, "Cheat sequence longer than supported by "
                                  "Vanilla dehacked");
